convert EnergySource to an interface and create 3 energy sources
